[CS-4932]:Migrate useAccProfile to ts
